chore: more clean up and error handling in the wasm client
